Rename scollIntoViewWithOffset to scrollIntoViewWithOffset
