+++ packages/core/src/components/Cell/utils/scrollIntoViewWithOffset.ts
